fix badger duplicate values
