+++ database/sqlite.go
@@ -4,76 +4,12 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/dgraph-io/badger/v4"
 	"github.com/jmoiron/sqlx"
 	_ "github.com/mattn/go-sqlite3"
 )
 
-type Handler interface {
-	Insert(key, val string) error
-	Get(key string) (string, error)
-	Close() error
-}
-
 type SqliteHandler struct {
 	db *sqlx.DB
-}
-
-type BadgerHandler struct {
-	db *badger.DB
-}
-
-func (d BadgerHandler) Close() error {
-	return d.db.Close()
-}
-
-func (d BadgerHandler) Insert(key, val string) error {
-	prevVal, err := d.Get(key)
-	var newVal string
-	// previous value does not exist
-	if err != nil {
-		newVal = val
-	} else {
-		newVal = prevVal + "," + val
-	}
-
-	return d.db.Update(func(txn *badger.Txn) error {
-		if len(key) == 0 {
-			return nil
-		}
-		return txn.Set([]byte(key), []byte(newVal))
-	})
-}
-
-func (d BadgerHandler) Get(key string) (string, error) {
-	var value []byte
-	err := d.db.View(func(txn *badger.Txn) error {
-		item, err := txn.Get([]byte(key))
-		if err != nil {
-			return fmt.Errorf("getting value: %w", err)
-		}
-		item.Value(func(val []byte) error {
-			value = append([]byte{}, val...)
-			return nil
-		})
-		return nil
-	})
-	if err != nil {
-		return string(value), err
-	}
-	return string(value), nil
-}
-
-func NewBadger(dataSourceName string) (*BadgerHandler, error) {
-	db, err := badger.Open(badger.DefaultOptions(dataSourceName))
-	if err != nil {
-		return nil, fmt.Errorf("opening db: %w", err)
-	}
-
-	handler := &BadgerHandler{
-		db: db,
-	}
-	return handler, nil
 }
 
 func NewSqlite(dataSouceName string) (*SqliteHandler, error) {
